修正函数返回值类型错误

+++ unipoint.cpp
@@ -38,7 +38,7 @@ Point::Point(double x, double y) : UniPoint(x, y)
 }
 
 
-double Point::GetClusterID(void)
+int Point::GetClusterID(void)
 {
     return cluster_id;
 }
